+++ test/FastaToAssembly_unit_test.py
@@ -5,16 +5,14 @@ Integration tests are in the server test file.
 
 import os
 import uuid
-
 from pathlib import Path
-from pytest import raises
-from typing import Optional, Callable
+from typing import Callable, Optional
 from unittest.mock import create_autospec
 
 from AssemblyUtil.FastaToAssembly import FastaToAssembly
-from installed_clients.DataFileUtilClient import DataFileUtil
 from conftest import assert_exception_correct
-
+from installed_clients.DataFileUtilClient import DataFileUtil
+from pytest import raises
 
 # TODO Add more unit tests when changing things until entire file is covered by unit tests
 
@@ -31,13 +29,13 @@ def _set_up_mocks(
     return fta, dfu
 
 
-def _run_test_spec_fail(test_spec, mass=False, print_spec=False):
+def _run_test_spec_fail(test_spec, max_cumsize=None, mass=False, print_spec=False):
     fta, _ = _set_up_mocks()
     for err, params in test_spec:
         if print_spec:
             print(f"spec:\n{params}")
         if mass:
-            _run_test_mass_fail(fta, params, err)
+            _run_test_mass_fail(fta, params, max_cumsize, err)
         else:
             _run_test_fail(fta, params, err)
 
@@ -48,9 +46,9 @@ def _run_test_fail(fta, params, expected):
         assert_exception_correct(got.value, ValueError(expected))
 
     
-def _run_test_mass_fail(fta, params, expected):
+def _run_test_mass_fail(fta, params, max_cumsize, expected):
         with raises(Exception) as got:
-            fta.import_fasta_mass(params)
+            fta.import_fasta_mass(params, 2.5, 10, max_cumsize, parallelize=False)
         assert_exception_correct(got.value, ValueError(expected))
 
 
@@ -222,7 +220,7 @@ def _test_import_fasta_mass_file(tmp_path, params_root):
             }
         ]
     })
-    res = fta.import_fasta_mass(params_root)
+    res = fta.import_fasta_mass(params_root, parallelize=False)
     assert res == [
         {'upa': '42/4/75', 'filtered_input': None},
         {'upa': '42/7/1', 'filtered_input': None}
@@ -422,7 +420,7 @@ def test_import_fasta_mass_blobstore_min_contig_length(tmp_path):
             {'node': 'fake_id_1', 'assembly_name': 'foo1'},
             {'node': 'fake_id_2', 'assembly_name': 'foo2'}
         ]
-    })
+    }, parallelize=False)
     assert res == [
         {'upa': '42/3/75', 'filtered_input': str(dir1 / 'f1.blobstore.fasta.filtered.fa')},
         {'upa': '42/6/1', 'filtered_input': str(dir2 / 'f2.blobstore.fasta.filtered.fa')}
@@ -625,3 +623,13 @@ def test_import_fasta_mass_fail_min_contig_length():
           _update(b, {'min_contig_length': {}})),
     ]
     _run_test_spec_fail(test_spec, mass=True)
+
+
+def test_import_fasta_mass_fail_invalid_max_cumsize():
+    b = {"workspace_id": 1, "inputs": [{"file": "b", "assembly_name": "x"}]}
+    test_spec1 = [("max_cumsize must be an integer or decimal", b)]
+    test_spec2 = [("max_cumsize must be > 0", b)]
+    test_spec3 = [(f"max_cumsize must be <= {1024 * 1024 * 1024 * 0.95}", b)]
+    _run_test_spec_fail(test_spec1, max_cumsize="9999", mass=True)
+    _run_test_spec_fail(test_spec2, max_cumsize=-1, mass=True)
+    _run_test_spec_fail(test_spec3, max_cumsize=1024 * 1024 * 1024 * 1024, mass=True)